API

API

Public

EconJobMarket.EconJobMarket — Module.

EconJobMarket

An interface with the EconJobMarket API.

source

EconJobMarket.fetch_ads — Method.

fetch_ads(filepath::AbstractString; delim::Union{Char,AbstractString} = '\t')

Writes a text file with the latest job opening advertisements from EJM.

Example

julia> data = joinpath(dirname(@__DIR__), "data", "ads.tsv") |>
       fetch_ads |>
       CSV.read;

julia> Tables.schema(data)
Tables.Schema:
 :posid                              Int64
 :oid                                Int64
 :adtitle                            String
 :position_type_id                   String
 :position_type                      String
 :adtext                             String
 :startdate                          Date
 :enddate                            Date
 :country_code                       String
 :position_country                   String
 :department                         String
 :shortname                          String
 :address                            Missing
 :longitude                          Float64
 :latitude                           Float64
 ⋮
 :business_economics                 Bool
 :computational_economics            Bool
 :political_economy                  Bool
 :accounting                         Bool
 :decision_sciences                  Bool
 :insurance                          Bool
 :management_general                 Bool
 :management_health_care             Bool
 :management_information_technology  Bool
 :marketing                          Bool
 :operations_research                Bool
 :organizational_behavior            Bool
 :real_estate                        Bool
 :statistics                         Bool

source

Private

EconJobMarket.CATEGORIES — Constant.

CATEGORIES::Vector{String}

Job opening advertisements categories.

julia> EconJobMarket.CATEGORIES
34-element Array{String,1}:
 "Development; Growth"
 "Econometrics"
 "Economic History"
 "Environmental; Ag. Econ."
 "Experimental Economics"
 "Finance"
 "Industrial Organization"
 "International Finance/Macro"
 "International Trade"
 "Labor; Demographic Economics"
 "Law and Economics"
 "Macroeconomics; Monetary"
 "Microeconomics"
 "Public Economics"
 "Theory"
 "Behavioral Economics"
 ⋮
 "Urban; Rural; Regional Economics"
 "Health; Education; Welfare"
 "Business Economics"
 "Computational Economics"
 "Political Economy"
 "Accounting"
 "Decision Sciences"
 "Insurance"
 "Management, General"
 "Management, Health Care"
 "Management, Information Technology"
 "Marketing"
 "Operations Research"
 "Organizational Behavior"
 "Real Estate"
 "Statistics"

source

EconJobMarket.CATEGORIES_URL — Constant.

CATEGORIES_URL::String

URL for accessing the public data for job opening categories in JSON format from EconJobMarket.

source

EconJobMarket.OPENINGS_URL — Constant.

OPENINGS_URL::String

URL for accessing the public data for job opening advertisements in JSON format from EconJobMarket.

source

EconJobMarket.categories_indicator — Method.

categories_indicator(obj::AbstractString)::BitVector

Given a string, it provides indicators for whether the job ad listed each opening category.

Example

julia> EconJobMarket.categories_indicator("2,3")
34-element BitArray{1}:
 0
 1
 1
 0
 0
 0
 0
 0
 0
 0
 ⋮
 0
 0
 0
 0
 0
 0
 0
 0
 0

source

EconJobMarket.clean_name — Method.

clean_name(obj)::Symbol

Normalize the value such that it qualifies as a lowercase valid identifier.

Example

julia> EconJobMarket.clean_name(Symbol("3This  is a Cat:egory!4cΩ."))
:_3this_is_a_cat_egory!4cΩ_

source

EconJobMarket.parse_opening — Method.

parse_opening(node)::NamedTuple

For each opening, it parses the information.

source